centralized-sampling-tests/sample-apps/python-flask/requirements.txt
boto3
certifi
chardet
Flask
idna
requests
urllib3
Werkzeug==3.0.1
flask-sqlalchemy==2.5.1
SQLAlchemy==1.4
aws_xray_sdk==2.6.0
opentelemetry-api
opentelemetry-sdk
